Sockets ported (shit-code)
